Removed old interceptors
[withErrorHandler.js]
If we want use this hoc globally we must remove old interceptors.

This will just ensure that whenever we don't need e.g. BurgerBuilder component anymore
that we clean up the interceptors which we attached due to using withErrorHandler
in the burgerBuilder so that if we reuse withErrorHandler in our application, we don't
create more and more interceptors with the old ones living on.